+++ NEWS.md
@@ -96,6 +96,7 @@ Development
 * Refactored Builder specs generation using Webpack (#11698)
 * Update cartodb.js to use multiple subdomains.
 * Update tangram to use subdomains.
+* New dashboard notifications added (#11807).
 
 ### Bug fixes
 * Fixed layers order when creating a new layer dragging from a compound analysis (#11827)